GEN AI/Kafka-Snowflake-Assistant/requirements.txt
streamlit
snowflake-snowpark
confluent_kafka
snowflake-connector-python
pandas
json
os
time
google-generativeai